+++ .firebase/hosting.cHVibGlj.cache
@@ -1,0 +1,27 @@
+index.html,1725409802039,ca504ff1ff651c76036c42701a855d246f48485888dc18ab1788cf98e20f18ca
+index.css,1719947794000,fd35484219bccf2a80b107248eb6d4d066247e5048083d4744b2c9403ad8aa12
+bd.js,1725289577963,332cc15335614e7c632942452a4c40639086d05ce19d0b62c9b789efc0a032c3
+404.html,1725406198033,762bf484ba67404bd1a3b181546ea28d60dfddf18e9dd4795d8d25bcf3c1a890
+Mestre/js/script.js,1719947794000,4cf1becdf58690c21a4a894b7b46f00520f9bda4c088dc0a6d54725374b3ec12
+Mestre/js/magias.js,1720807163097,e326fc4f847a61a47ccce3b10de5d187e215920f47c1f1fdb5c465b023045384
+Mestre/js/itemlist.js,1720528601425,44d1cd435eff84c7c3863777fbd41038b54d8186edb99ebbea754049cf88e89a
+Mestre/html/magias.html,1719947794000,fb6202f36e6b8828ea1d4031b9b100d6c79991fe2799221a0fa444abd28e5bc7
+Mestre/html/itemlist.html,1719947794000,051da0399383f7d8d0437ab0c6574c8d9e6de471ce49c7de303d8a8094bf02c2
+Mestre/html/cadastro_itens.html,1720203201030,e955a68246dc5ec38f00ee3cfde0cd8647d1214ec5cc474a62cbc4561e3b10d4
+Mestre/css/magias.css,1719947794000,1691ec97153986db72c9bd1c135b99810a885486ca145576cc4b5c38069ab790
+Mestre/css/itemlist.css,1719947794000,70baa1fcf98d6c5ee63748861c79eb2db3a61a753e350b554c4ee28a58c939fb
+Mestre/css/cadastro_itens.css,1719947794000,e4ca034611c3638d76a87ec5f8a865b5bfb18ea8c1aaabfaff90f8af271f0de1
+Jogador/jogador.html,1725454990296,a90105fb559277e6b34fecdb317f41141a98fa25a55afff4748fec52b5d14567
+Jogador/jogador.css,1725455015299,0959a34f14a26a9b938569438a64f5af49b73e86e5e8bd5fc96c215b70836d48
+Jogador/js/modal.js,1725297744565,2af7b0bd6112a7eac9a61c502716ceafaecf06ce49ef46ff64a580ed6f4ac830
+Jogador/js/jogadorNotas.js,1725390690641,865b5d6f4719f63011c67209bc3d4a81b9cfc5ab889134ac479b35c4de3837ae
+Jogador/js/jogadorMagia.js,1725454656236,c6c857babf080e6df67f458c320a04e5ccc73672897995f2d92ffcf04041bab6
+Jogador/js/jogadorLogin.js,1725037893898,46b50c321b39e89a491b6727a01628c34245605a30beb3e7414c5e01cff90e6e
+Jogador/js/jogadorItem.js,1725451078821,7b4ed59f54e523a4ba4b98f3817ba00fce5cb8a01a18849df79bed002d19f828
+Jogador/js/jogadorFicha.js,1725451107953,ae4b37139faf19def554dbbe8dc902187547b20257c18d475ab3b866fe3516b1
+img/slogan.jpg,1719947794000,49f1becc8635a092559d3c14056ea1fe3bf1420ada0d634b0e9162bcd94f2a7a
+img/mestre.jpg,1719947794000,9cda238f18c65897d6fc351c6d9450bb419c7c44833ead87e524d4c80f368f6c
+img/logo.png,1719947794000,6e8a8c99893123eec91beb3040431c38840cd41ae4a4d503f8fd2e3e18f62a34
+img/jogadores.jpg,1719947794000,d0e79b1b695d6e7aee8ba7de9c07dc24a5649ee7b8a379be1299efda8ea13436
+img/img.jpg,1719947794000,3ec8878c9c3e7fc8d3104fa856b3b2d2b7d1d2e36a956ed6e56d20ed13fab106
+img/bg.jpg,1719947794000,930d06aae689960861490be1d2de9f5ac0980a553012affa6b68bcaafa016924
