+++ .firebaserc
@@ -1,0 +1,5 @@
+{
+  "projects": {
+    "default": "arcanum-bd4b1"
+  }
+}
