--- index.html
@@ -1,57 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Dragon's Grimoire</title>
-    <link rel="stylesheet" href="index.css">
-</head>
-<body>
-    <header>
-        <div class="logo-container">
-            <img src="img/logo.png" alt="Arcanum Logo">
-        </div>
-    </header>
-
-    <main>
-
-        <section class="about-section">
-            <div class="slogan">
-                <div class="div_slogan">
-                    <h2>Dragon's Grimoire?</h2>
-                <p>Lorem, ipsum dolor sit amet consectetur adipisicing elit. Necessitatibus asperiores eligendi dolorem quisquam voluptatem quos dolorum odit explicabo quas consequatur, quod in ea ut, laborum veniam laboriosam molestias labore. Sed..</p>
-                </div>
-               <div >
-                <img class="img_slogan" src="img/slogan.jpg" alt="">
-               </div>
-                
-            </div>
-        </section>
-        <div class="tarja">
-            <p>asdasdasdasdasdasdasda</p>
-        </div>
-        <section class="user-options">
-            <div class="user-option">
-                <img  class="img_jogador" src="img/jogadores.jpg" alt="Jogador">
-                <a href="./Jogador/jogador.html"><button class="player-button">Jogador</button></a>
-                <!-- <p class="p_jogador">Crie seu personagem,liste seus itens e suas magias.</p> -->
-            </div>
-
-            <div class="user-option">
-                <img   class="img_mestre" src="img/mestre.jpg" alt="Mestre">
-                <a href="./Mestre/html/cadastro_itens.html"><button class="dm-button">Mestre</button></a>
-                <!-- <p>Crie e gerencie itens, magias e anotações para suas campanhas.</p> -->
-            </div>
-        </section>
-
-        <section class="project-description">
-            <h2>Sobre o Projeto</h2>
-            <p>Texto sobre o projeto, sua finalidade e como funciona. Espaço para informações adicionais.</p>
-        </section>
-    </main>
-
-    <footer>
-        <p>&copy; 2024 Seu Nome. Todos os direitos reservados.</p>
-    </footer>
-</body>
-</html>
