+++ public/Jogador/js/jogadorNotas.js
