+++ public/Mestre/js/magias.js
