+++ public/index.html
@@ -1,0 +1,116 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title>Welcome to Firebase Hosting</title>
+
+    <!-- update the version number as needed -->
+    <script defer src="/__/firebase/10.13.1/firebase-app-compat.js"></script>
+    <!-- include only the Firebase features as you need -->
+    <script defer src="/__/firebase/10.13.1/firebase-auth-compat.js"></script>
+    <script defer src="/__/firebase/10.13.1/firebase-database-compat.js"></script>
+    <script defer src="/__/firebase/10.13.1/firebase-firestore-compat.js"></script>
+    <script defer src="/__/firebase/10.13.1/firebase-functions-compat.js"></script>
+    <script defer src="/__/firebase/10.13.1/firebase-messaging-compat.js"></script>
+    <script defer src="/__/firebase/10.13.1/firebase-storage-compat.js"></script>
+    <script defer src="/__/firebase/10.13.1/firebase-analytics-compat.js"></script>
+    <script defer src="/__/firebase/10.13.1/firebase-remote-config-compat.js"></script>
+    <script defer src="/__/firebase/10.13.1/firebase-performance-compat.js"></script>
+    <!-- 
+      initialize the SDK after all desired features are loaded, set useEmulator to false
+      to avoid connecting the SDK to running emulators.
+    -->
+    <script defer src="/__/firebase/init.js?useEmulator=true"></script>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet"
+    integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
+        <link rel="stylesheet" type="text/css" href="index.css">
+  </head>
+  <body>
+    <header>
+        <div class="logo-container">
+            <img src="img/logo.png" alt="Arcanum Logo">
+        </div>
+    </header>
+
+    <main>
+
+        <section class="about-section">
+            <div class="slogan">
+                <div class="div_slogan">
+                    <h2>Dragon's Grimoire?</h2>
+                <p>Lorem, ipsum dolor sit amet consectetur adipisicing elit. Necessitatibus asperiores eligendi dolorem quisquam voluptatem quos dolorum odit explicabo quas consequatur, quod in ea ut, laborum veniam laboriosam molestias labore. Sed..</p>
+                </div>
+               <div >
+                <img class="img_slogan" src="img/slogan.jpg" alt="">
+               </div>
+                
+            </div>
+        </section>
+        <div class="tarja">
+            <p>asdasdasdasdasdasdasda</p>
+        </div>
+        <section class="user-options">
+            <div class="user-option">
+                <img  class="img_jogador" src="img/jogadores.jpg" alt="Jogador">
+                <a href="./Jogador/jogador.html"><button class="player-button">Jogador</button></a>
+                <!-- <p class="p_jogador">Crie seu personagem,liste seus itens e suas magias.</p> -->
+            </div>
+
+            <div class="user-option">
+                <img   class="img_mestre" src="img/mestre.jpg" alt="Mestre">
+                <a href="./Mestre/html/cadastro_itens.html"><button class="dm-button">Mestre</button></a>
+                <!-- <p>Crie e gerencie itens, magias e anotações para suas campanhas.</p> -->
+            </div>
+        </section>
+
+        <section class="project-description">
+            <h2>Sobre o Projeto</h2>
+            <p>Texto sobre o projeto, sua finalidade e como funciona. Espaço para informações adicionais.</p>
+        </section>
+    </main>
+
+    <footer>
+        <p>&copy; 2024 Seu Nome. Todos os direitos reservados.</p>
+    </footer>
+
+    <script>
+      document.addEventListener('DOMContentLoaded', function() {
+        const loadEl = document.querySelector('#load');
+        // // 🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥
+        // // The Firebase SDK is initialized and available here!
+        //
+        // firebase.auth().onAuthStateChanged(user => { });
+        // firebase.database().ref('/path/to/ref').on('value', snapshot => { });
+        // firebase.firestore().doc('/foo/bar').get().then(() => { });
+        // firebase.functions().httpsCallable('yourFunction')().then(() => { });
+        // firebase.messaging().requestPermission().then(() => { });
+        // firebase.storage().ref('/path/to/ref').getDownloadURL().then(() => { });
+        // firebase.analytics(); // call to activate
+        // firebase.analytics().logEvent('tutorial_completed');
+        // firebase.performance(); // call to activate
+        //
+        // // 🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥🔥
+
+        try {
+          let app = firebase.app();
+          let features = [
+            'auth', 
+            'database', 
+            'firestore',
+            'functions',
+            'messaging', 
+            'storage', 
+            'analytics', 
+            'remoteConfig',
+            'performance',
+          ].filter(feature => typeof app[feature] === 'function');
+          loadEl.textContent = `Firebase SDK loaded with ${features.join(', ')}`;
+        } catch (e) {
+          console.error(e);
+          loadEl.textContent = 'Error loading the Firebase SDK, check the console.';
+        }
+      });
+    </script>
+  </body>
+</html>
